requirements.txt
h5py==3.1.0
matplotlib==3.3.0
networkx==2.5.1
numpy==1.19.5
plotly==4.12.0
scipy==1.7.0
six==1.15.0
tensorflow-gpu==2.6.0
wandb==0.10.33